Backend/Netra_Backend/netra_backend/services/OD_service/requirements.txt
# OD Service Requirements
# Python dependencies for the Orbit Determination Service

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
psycopg2-binary==2.9.9

# Scheduler
APScheduler==3.10.4
pytz==2024.1

# Additional utilities (optional but recommended)
python-multipart==0.0.9  # For form data handling in FastAPI
pydantic==2.5.3  # Data validation (comes with FastAPI but pinning version)

# TLE to WGS84 conversion
sgp4==2.23  # Satellite orbit propagation
numpy==1.26.4 # Numerical operations
